Added version 1.2 APK.
This version includes an additional arc for deployed flaps. Additionally the previous method for defining a "warning" angle has gone away. The app will automatically add a yellow arc 30% of the distance from 0 degrees to the defined "danger" angle (e.g. the critical angle of attack). This is true for both arcs - e.g. flaps and no flaps.